fix(zoom): change all zoom level types from Int32 to Double/Float
TypeScript spec files (NativeModule + NativeView codegen specs): all
zoom-related fields changed from Int32 to Double:
- NativeMapContainer.ModuleParams: zoomLevel, zoomMin, zoomMax
- NativeMapContainer.GetPositionResponse: zoomLevel
- MapsforgeVtmViewNativeComponent.MapViewProps: zoomLevel, zoomMin, zoomMax
- NativeLayerMapsforge: enabledZoomMin, enabledZoomMax (ModuleParams,
  CreateSubLayerParams, UpdateEnabledZoomMinMaxParams)
- NativeLayerHillshading: zoomMin, zoomMax, enabledZoomMin, enabledZoomMax
- NativeLayerBitmapTile: zoomMin, zoomMax, enabledZoomMin, enabledZoomMax
- NativeLayerMBTilesBitmap: enabledZoomMin, enabledZoomMax,
  LayerMBTilesBitmapResponse.zoomMin, zoomMax
- NativeLayerPath: GeometryStyle.scalingZoomLevel
- NativeLayerPathJts: GeometryStyleJts.scalingZoomLevel
- NativeLayerShape: ShapeStyle.scalingZoomLevel

TSX components: removed Math.round() wrapping from all zoom prop values
since they are no longer constrained to integers.

Java: MapsforgeVtmView stores zoom as float instead of int;
MapsforgeVtmViewManager setters accept double (matching codegen output);
params.getInt() replaced with (int) params.getDouble() for zoom fields;
MapFragment casts float→int at the vtm API boundary
(getZoomLevel()/setMinZoomLevel/setMaxZoomLevel, which remain int in vtm).

+++ android/src/main/java/com/jhotadhari/reactnative/mapsforge/vtm/LayerZoomBoundsHelper.java
@@ -28,8 +28,8 @@ public class LayerZoomBoundsHelper extends LayerHelper {
 		if ( null == mapView ) { return null; }
 
 		// Get params, assign defaults.
-		int enabledZoomMin = Utils.rMapHasKey( params, "enabledZoomMin" ) ? params.getInt( "enabledZoomMin" ) : (int) module.getConstants().get( "enabledZoomMin" );
-		int enabledZoomMax = Utils.rMapHasKey( params, "enabledZoomMax" ) ? params.getInt( "enabledZoomMax" ) : (int) module.getConstants().get( "enabledZoomMax" );
+		int enabledZoomMin = Utils.rMapHasKey( params, "enabledZoomMin" ) ? (int) params.getDouble( "enabledZoomMin" ) : (int) module.getConstants().get( "enabledZoomMin" );
+		int enabledZoomMax = Utils.rMapHasKey( params, "enabledZoomMax" ) ? (int) params.getDouble( "enabledZoomMax" ) : (int) module.getConstants().get( "enabledZoomMax" );
 
 		updateEnabled(
 			layer,
@@ -55,8 +55,8 @@ public class LayerZoomBoundsHelper extends LayerHelper {
 			}
 
 			// Get params, assign defaults.
-			int enabledZoomMin = Utils.rMapHasKey( params, "enabledZoomMin" ) ? params.getInt( "enabledZoomMin" ) : (int) module.getConstants().get( "enabledZoomMin"  );
-			int enabledZoomMax = Utils.rMapHasKey( params, "enabledZoomMax" ) ? params.getInt( "enabledZoomMax" ) : (int) module.getConstants().get( "enabledZoomMax"  );
+			int enabledZoomMin = Utils.rMapHasKey( params, "enabledZoomMin" ) ? (int) params.getDouble( "enabledZoomMin" ) : (int) module.getConstants().get( "enabledZoomMin"  );
+			int enabledZoomMax = Utils.rMapHasKey( params, "enabledZoomMax" ) ? (int) params.getDouble( "enabledZoomMax" ) : (int) module.getConstants().get( "enabledZoomMax"  );
 
 			updateUpdateListener(
 				params.getInt( "nativeNodeHandle" ),
