+++ src/NativeModules/NativeLayerShape.ts
@@ -27,7 +27,7 @@ export type ShapeStyle = {
 	fillColor?: `#${string}`;
 	fillAlpha?: Double;
 	buffer?: Double;
-	scalingZoomLevel?: Int32;
+	scalingZoomLevel?: Double;
 	cap?: 'SQUARE' | 'ROUND' | 'BUTT';
 	fixed?: boolean;
 	strokeIncrease?: Double;
@@ -137,7 +137,7 @@ interface UpdateShapeParams {
 		fillColor?: string;
 		fillAlpha?: Double;
 		buffer?: Double;
-		scalingZoomLevel?: Int32;
+		scalingZoomLevel?: Double;
 		cap?: string;
 		fixed?: boolean;
 		strokeIncrease?: Double;
